+++ cmd/exec_cmd.go
@@ -86,13 +86,13 @@ func (cc *ExecCmd) Run(ctx *RunContext) error {
 	fmt.Printf("Please use `exit` or `Ctrl-D` to quit.\n")
 
 	c := NewTagsCompleter(ctx, sso, execCmd)
+	opts := ctx.Settings.DefaultOptions(c.ExitChecker)
+	opts = append(opts, ctx.Settings.GetColorOptions()...)
+
 	p := prompt.New(
 		c.Executor,
 		c.Complete,
-		prompt.OptionPrefix("> "),
-		prompt.OptionSetExitCheckerOnInput(c.ExitChecker),
-		prompt.OptionCompletionOnDown(),
-		prompt.OptionShowCompletionAtStart(),
+		opts...,
 	)
 	p.Run()
 	return nil